+++ spec/fixtures/lib/test_i18n_plugin.rb
@@ -10,3 +10,5 @@ module TestI18nPlugin
     puts "Hello, #{opts[:name]}"
   end
 end
+
+::I18n::Tasks.add_commands TestI18nPlugin